Lectures/30-Project_Python/requirements.txt
rich
requests
fastapi
uvicorn
pydantic
InquirerPyrich
